refactor: split markup into components
