+++ src/entities/achievements/index.ts
@@ -1,0 +1,2 @@
+export { AchievementsOverview } from "./overview";
+export { AchievementsPopup } from "./popup";
