+++ src/entities/bone-trader/index.ts
@@ -1,0 +1,1 @@
+export { BoneTrader } from "./trader";
