+++ src/entities/heirlooms/index.ts
@@ -1,0 +1,2 @@
+export { HeirloomsOverview } from "./overview";
+export { HeirloomsPopup } from "./popup";
