+++ src/entities/main-screen/index.ts
@@ -1,0 +1,1 @@
+export { MainScreen } from "./screen";
