+++ src/entities/mutations/index.ts
@@ -1,0 +1,1 @@
+export { MutationTree } from "./tree";
