+++ src/entities/mutations/tree.tsx
@@ -1,0 +1,9 @@
+export function MutationTree() {
+  return (
+    <div
+      id="mutTreeWrapper"
+      style={{ display: "none" }}
+      className="noselect"
+    ></div>
+  );
+}
