+++ src/entities/offline/index.ts
@@ -1,0 +1,1 @@
+export { OfflineWrapper } from "./wrapper";
