+++ src/entities/player-spire/index.ts
@@ -1,0 +1,1 @@
+export { PlayerSpire } from "./spire";
