+++ src/entities/player-spire/spire.tsx
@@ -1,0 +1,20 @@
+export function PlayerSpire() {
+  return (
+    <div id="playerSpirePopout" style={{ display: "none" }}>
+      <div id="playerSpireSpirePanel">
+        <div id="playerSpireSpireSpirePanel" className="noselect">
+          &nbsp;
+        </div>
+        <div id="floatingCombatText"></div>
+        &nbsp;
+      </div>
+
+      <div id="playerSpireSmallPanel" className="niceScroll">
+        &nbsp;
+      </div>
+      <div id="playerSpireInfoPanel" className="niceScroll">
+        &nbsp;
+      </div>
+    </div>
+  );
+}
