+++ src/entities/portal/index.ts
@@ -1,0 +1,1 @@
+export { Portal } from "./portal";
