+++ src/entities/stats/index.ts
@@ -1,0 +1,1 @@
+export { StatsOverview } from "./overview";
