+++ src/entities/tooltip/index.ts
@@ -1,0 +1,2 @@
+export { Tooltip } from "./tooltip";
+export { Tooltip2 } from "./tooltip-2";
