+++ src/entities/tooltip/tooltip-2.tsx
@@ -1,0 +1,15 @@
+export function Tooltip2() {
+  return (
+    <div
+      id="tooltipDiv2"
+      className="tooltipExtraNone"
+      style={{ display: "none" }}
+    >
+      <div id="tipTitle2"></div>
+      <hr />
+      <div id="tipText2"></div>
+      <hr />
+      <div id="tipCost2"></div>
+    </div>
+  );
+}
