+++ src/entities/tutorial/index.ts
@@ -1,0 +1,1 @@
+export { Tutorial } from "./tutorial";
