+++ src/styles/global.scss
@@ -1,3 +1,2 @@
 @use "./sass/variables.scss" as *;
 @use "./variables.scss";
-// @use "./original/style.scss"
